% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelewoc.R
\name{CompleteModelSeason}
\alias{CompleteModelSeason}
\title{Function modeling the behavior of the agro-ecosystem during nbnT years}
\usage{
CompleteModelSeason(nbnT, np, nf, parms)
}
\arguments{
\item{nbnT}{duration of simulation (in years)}

\item{np}{length of a cycle (in years)}

\item{nf}{duration of fallow (in years) with nf < np}

\item{parms}{num}
}
\value{
list
}
\description{
Function modeling the behavior of the agro-ecosystem during nbnT years
}
\details{
It take as argument the numbers of years to simulate (nbnT),
   the number of parcels in bush ring (np),
   the number of parcels in Fllws in bush ring (nf)
   the list of parms (see the end of document)
}
\examples{
# not run
}
