% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycleData}
\alias{cycleData}
\title{Outputs of several simulations}
\format{A data frame with 368 rows and 47 variables:
\describe{
\item{\code{aireB}}{a numeric vector}
\item{\code{kapa}}{a numeric vector}
\item{\code{np}}{a numeric vector}
\item{\code{nf}}{a numeric vector}
\item{\code{OM_Case}}{a numeric vector}
\item{\code{OM_Gdnt}}{a numeric vector}
\item{\code{OM_Fllw}}{a numeric vector}
\item{\code{OD_Case}}{a numeric vector}
\item{\code{OD_Gdnt}}{a numeric vector}
\item{\code{OD_Fllw}}{a numeric vector}
\item{\code{IP_Gdnt}}{a numeric vector}
\item{\code{IP_Fllw}}{a numeric vector}
\item{\code{Mi_Case}}{a numeric vector}
\item{\code{Mi_Fllw}}{a numeric vector}
\item{\code{Mi_Gdnt}}{a numeric vector}
\item{\code{P1}}{a numeric vector}
\item{\code{D1}}{a numeric vector}
\item{\code{M1}}{a numeric vector}
\item{\code{P10}}{a numeric vector}
\item{\code{D10}}{a numeric vector}
\item{\code{M10}}{a numeric vector}
\item{\code{PC}}{a numeric vector}
\item{\code{DC}}{a numeric vector}
\item{\code{MC}}{a numeric vector}
\item{\code{food_in_Fllw}}{a numeric vector}
\item{\code{food_in_Gdnt}}{a numeric vector}
\item{\code{food_in_case}}{a numeric vector}
\item{\code{food_in_sava}}{a numeric vector}
\item{\code{food_ou_Fllw}}{a numeric vector}
\item{\code{food_ou_Gdnt}}{a numeric vector}
\item{\code{food_ou_case}}{a numeric vector}
\item{\code{harvest}}{a numeric vector}
\item{\code{compound_harvest}}{a numeric vector}
\item{\code{export_harvest}}{a numeric vector}
\item{\code{bush_harvest}}{a numeric vector}
\item{\code{frac}}{a numeric vector}
\item{\code{total_harv}}{a numeric vector}
\item{\code{prod_bush}}{a numeric vector}
\item{\code{prod_comp}}{a numeric vector}
\item{\code{prod_tot}}{a numeric vector}
\item{\code{rend_comp}}{a numeric vector}
\item{\code{rend_bush}}{a numeric vector}
\item{\code{rend_tot}}{a numeric vector}
\item{\code{Livestock}}{a factor with levels \code{0} \code{410}}
\item{\code{airB}}{a factor with levels \code{75 \%}}
\item{\code{testbush}}{a numeric vector}
\item{\code{id}}{a character vector}
}}
\usage{
cycleData
}
\description{
A dataset containing Outputs of several simulations obtained with the parameters parms in which only change
the herbivory pressure (livestock is 0 or 410)
the length of cycle (10, 20, 50 and 100 years long cycle)
and the proportion of fallow during cycle.
}
\keyword{datasets}
