% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelewoc.R
\name{DrySeason}
\alias{DrySeason}
\title{systems of differential equations modeling the behavior of the agro-ecosystem during the dry season}
\usage{
DrySeason(t, x, parametre)
}
\arguments{
\item{t}{numeric vector, duration of simulation}

\item{x}{numeric vector, initial  condition of state variables}

\item{parametre}{list, list of initialization of parameters 'models}
}
\value{
list, contains dx
}
\description{
systems of differential equations modeling the behavior of the agro-ecosystem during the dry season
}
\details{
xxxx
}
\examples{
# not run
# test<-DrySeason(t=,x=,parametre=)
}
