% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelewoc.R
\name{function_AS}
\alias{function_AS}
\title{function_AS: cette fonction renvoie des sorties ponctuelles du modele definies a la fin du code}
\usage{
function_AS(param_AS, npp, nff, nbnTt)
}
\arguments{
\item{param_AS}{parametres dont on fait l'analyse de sensibilité (matrix nbcol = nb de parametres, nbrow = nb de jeux de parametres)}

\item{npp}{longueur de rotation}

\item{nff}{durée en jachère (nff<npp)}

\item{nbnTt}{nombre d'années simulées}
}
\value{
sorties d'interet pour chaque jeu de parametres
}
\description{
function_AS: cette fonction renvoie des sorties ponctuelles du modele definies a la fin du code
}
\examples{
# not run
}
