% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelewoc.R
\name{GrowingSeason}
\alias{GrowingSeason}
\title{systems of differential equations modeling the behavior
of the agro-ecosystem during the growing season}
\usage{
GrowingSeason(t, x, parametre)
}
\arguments{
\item{t}{numeric vector, duration of simulation}

\item{x}{numeric vector, initial  condition of state variables}

\item{parametre}{list, list of initialization of parameters models}
}
\value{
list, contains vector and dataframe
}
\description{
systems of differential equations modeling the behavior
of the agro-ecosystem during the growing season
}
\details{
xxxx
}
\examples{
# not run
}
