% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parms}
\alias{parms}
\title{parameters}
\format{A list of 39 parameters for the ewoc model
\describe{
\item{\code{KC}}{a numeric}
\item{\code{KJ}}{a numeric}
\item{\code{KBmax}}{a numeric}
\item{\code{ipB}}{a numeric}
\item{\code{ipJ}}{a numeric}
\item{\code{bC}}{a numeric}
\item{\code{bB}}{a numeric}
\item{\code{bJ}}{a numeric}
\item{\code{ccC}}{a numeric}
\item{\code{ccB}}{a numeric}
\item{\code{ccJ}}{a numeric}
\item{\code{m}}{a numeric}
\item{\code{odC}}{a numeric}
\item{\code{odB}}{a numeric}
\item{\code{odJ}}{a numeric}
\item{\code{omC}}{a numeric}
\item{\code{omB}}{a numeric}
\item{\code{omJ}}{a numeric}
\item{\code{rho}}{a numeric}
\item{\code{kapa}}{a numeric}
\item{\code{fu}}{a numeric}
\item{\code{KK}}{a numeric}
\item{\code{lambdK}}{a numeric}
\item{\code{tau}}{a numeric}
\item{\code{co_case}}{a numeric}
\item{\code{co_chbu}}{a numeric}
\item{\code{co_fabu}}{a numeric}
\item{\code{aireB}}{a numeric}
\item{\code{aireT}}{a numeric}
\item{\code{im}}{a numeric}
\item{\code{id}}{a numeric}
\item{\code{lambd}}{a numeric}
\item{\code{gmC}}{a numeric}
\item{\code{gmB}}{a numeric}
\item{\code{exp}}{a numeric}
\item{\code{lgthss}}{a numeric}
\item{\code{r0}}{a numeric}
\item{\code{parKB}}{a numeric}
\item{\code{root}}{a numeric}
\item{\code{psi}}{a numeric}
\item{\code{zeta_dry}}{a numeric}
\item{\code{zeta_wet}}{a numeric}
\item{\code{gmax}}{a numeric}
}}
\usage{
parms
}
\description{
A list used in modelling
}
\keyword{datasets}
