% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphewoc.R
\name{plotLiveProd.cycle}
\alias{plotLiveProd.cycle}
\title{Production in function of cycle and cattle}
\usage{
plotLiveProd.cycle(datain, cbPalette, titleGraph, prod_chosed, airB_chosed)
}
\arguments{
\item{datain}{a dataframe}

\item{cbPalette}{chosen color (3)}

\item{titleGraph}{character title of graph}

\item{prod_chosed}{character production variable chosed}

\item{airB_chosed}{factor}
}
\value{
a ggplot2 graph
}
\description{
Production in function of cycle and cattle
}
\examples{
# not run
}
