% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphewoc.R
\name{plotRdtProd.cycle}
\alias{plotRdtProd.cycle}
\title{Rendement et production in function of cycle}
\usage{
plotRdtProd.cycle(datain, cbPalette, titleGraph, rend_chosed, prod_chosed)
}
\arguments{
\item{datain}{a dataframe}

\item{cbPalette}{color choosed (3)}

\item{titleGraph}{character title of graph}

\item{rend_chosed}{character rendement variable chosed}

\item{prod_chosed}{character production variable chosed}
}
\value{
a ggplot2 graph
}
\description{
Rendement et production in function of cycle
}
\examples{
# not run
}
