% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphewoc.R
\name{plotRdtProd.surface}
\alias{plotRdtProd.surface}
\title{Rendement et production in function of surface of compound and bush}
\usage{
plotRdtProd.surface(datain, cbPalette, titleGraph, livestock_chosed)
}
\arguments{
\item{datain}{a dataframe}

\item{cbPalette}{color choosed (3)}

\item{titleGraph}{character title of graph}

\item{livestock_chosed}{number}
}
\value{
a ggplot2 graph
}
\description{
Rendement et production in function of surface of compound and bush
}
\examples{
# not run
}
