% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphewoc.R
\name{plotStock}
\alias{plotStock}
\title{Stock of one simulation}
\usage{
plotStock(outputSimu, nbnT, np, title, cbPalette)
}
\arguments{
\item{outputSimu}{a dataframe from one simulation of model_complet_saison}

\item{nbnT}{number of year simulated in model_complet season (length of colomn time/365)}

\item{np}{number of parcelle of bush in simulation of modele_complet season}

\item{title}{character title of graph}

\item{cbPalette}{colour choosed (particular format, see Rcolorchart)}
}
\value{
a ggplot graph format pdf
}
\description{
Stock of one simulation
}
\examples{
# not run
}
