% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelewoc.R
\name{SurfaceVariation}
\alias{SurfaceVariation}
\title{CompleteModelSeason in function of different configuration of the agro-ecosystem
 (surface of Bush:Compound Ratio)}
\usage{
SurfaceVariation(np, nf, livestock, aireBush, parms, nbnT)
}
\arguments{
\item{np}{length of a cycle (in years)}

\item{nf}{duration of fallow (in years) with nf < np}

\item{livestock}{herbivory pressure (in Tropical livestock Units)}

\item{aireBush}{surface of bush ring}

\item{parms}{parms}

\item{nbnT}{duration of simulation (in years)}
}
\value{
dataframe usable by graphic function plotRdtProd.surface
}
\description{
CompleteModelSeason in function of different configuration of the agro-ecosystem
 (surface of Bush:Compound Ratio)
}
\examples{
# not run
}
